% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rf_metrics.R
\name{lfa_calculate_rf_metrics}
\alias{lfa_calculate_rf_metrics}
\title{Calculate Precision and Recall Metrics for Random Forest Classification}
\usage{
lfa_calculate_rf_metrics(conf_matrix)
}
\arguments{
\item{conf_matrix}{Confusion matrix obtained from a Random Forest classification.}
}
\value{
A data frame containing precision and recall metrics for each class.
}
\description{
This function calculates precision and recall metrics for each class based on the provided confusion matrix from a Random Forest classification.
}
\details{
The function calculates precision and recall metrics for each class based on the confusion matrix obtained from a Random Forest classification.
}
\examples{
# Example confusion matrix from a Random Forest classification
rf_cm <- table(predicted = c("A", "B", "A", "B"), actual = c("A", "A", "B", "B"))
# Calculate precision and recall metrics
rf_metrics_df <- lfa_calculate_rf_metrics(rf_cm)

}
\seealso{
\code{\link{lfa_precision_per_class}}, \code{\link{lfa_recall_per_class}}
}
