% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_flags.R
\name{lfa_check_flag}
\alias{lfa_check_flag}
\title{Check if a flag is set, indicating the completion of a specific process.}
\usage{
lfa_check_flag(flag_name)
}
\arguments{
\item{flag_name}{A character string specifying the name of the flag file. It should be a descriptive and unique identifier for the process being checked.}
}
\value{
A logical value indicating whether the flag is set (\code{TRUE}) or not (\code{FALSE}).
}
\description{
This function checks for the existence of a hidden flag file at a specified location within the working directory. If the flag file is found, a message is printed, and the function returns \code{TRUE} to indicate that the associated processing step has already been completed. If the flag file is not found, the function returns \code{FALSE}, indicating that further processing can proceed.
}
\examples{
# Check if the flag for a process named "data_processing" is set
lfa_check_flag("data_processing")
}
