% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chm.R
\name{lfa_chm}
\alias{lfa_chm}
\title{Create Canopy Height Model (CHM) from Lidar Data}
\usage{
lfa_chm(specie, area, res = 0.5, save_to_file = TRUE, overwrite = FALSE, ...)
}
\arguments{
\item{specie}{Character string indicating the species name.}

\item{area}{Character string indicating the specific area or location.}

\item{res}{Numeric value indicating the spatial resolution of the CHM. Default is 0.5.}

\item{save_to_file}{Logical. If TRUE, the generated CHM will be saved to a GeoTIFF file. Default is TRUE.}

\item{overwrite}{Logical. If TRUE, existing CHM file will be overwritten. Default is FALSE.}

\item{...}{Additional arguments to be passed to the underlying functions, such as lidR::catalog_map.}
}
\value{
A raster layer representing the Canopy Height Model (CHM).
}
\description{
This function generates a Canopy Height Model (CHM) from Lidar data using the lidR package.
}
\details{
The behavior of the function with different input parameters is as follows:
\itemize{
\item When a CHM file already exists at the specified path and \code{overwrite} is FALSE, the function loads the existing CHM and returns it.
\item If the CHM file does not exist or \code{overwrite} is TRUE, the function processes Lidar data using \code{lfa_rasterize_chunk} and creates a CHM.
\item The spatial resolution of the CHM can be controlled with the \code{res} parameter.
\item If \code{save_to_file} is TRUE, the generated CHM will be saved to a GeoTIFF file.
}
}
\examples{
# Generate CHM for a specific species and area
chm <- lfa_chm(specie = "ExampleSpecies", area = "ExampleArea", res = 1.0)

# Generate CHM and save it to a file
chm <- lfa_chm(specie = "ExampleSpecies", area = "ExampleArea", res = 1.0, save_to_file = TRUE)

}
\seealso{
\code{\link{lfa_read_area_as_catalog}}, \code{\link{catalog_map}}, \code{\link{terra::rast}}, \code{\link{terra::writeRaster}}
}
