% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sf_obj.R
\name{lfa_combine_sf_obj}
\alias{lfa_combine_sf_obj}
\title{Combine Spatial Feature Objects from Multiple GeoPackage Files}
\usage{
lfa_combine_sf_obj(paths, area_infos)
}
\arguments{
\item{paths}{A character vector containing file paths to GeoPackage files with neighbor information.}

\item{area_infos}{A data frame or list containing information about the corresponding detection areas, including "area" and "specie" columns.}
}
\value{
A combined sf object with additional columns for area and specie information.
}
\description{
This function reads spatial feature objects (sf) from multiple GeoPackage files and combines them into a single sf object.
Each GeoPackage file is assumed to contain neighbor information for a specific detection area, and the resulting sf object
includes additional columns indicating the corresponding area and species information.
}
\examples{
# Assuming paths and area_infos are defined
combined_sf <- lfa_combine_sf_obj(paths, area_infos)

# Print the combined sf object
print(combined_sf)

}
