% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_returns_per_tree.R
\name{lfa_count_returns_per_tree}
\alias{lfa_count_returns_per_tree}
\title{Count returns per tree for a given lidR catalog.}
\usage{
lfa_count_returns_per_tree(ctg)
}
\arguments{
\item{ctg}{A lidR catalog object containing LAS files to be processed.}
}
\value{
A data frame with columns for tree ID and the corresponding count of returns.
}
\description{
This function takes a lidR catalog as input and counts the returns per tree.
It uses the lidR package to read LAS files from the catalog and performs the counting
operation on each tree. The result is a data frame containing the counts of returns
for each unique tree ID within the lidR catalog.
}
\examples{
# Count returns per tree for a lidR catalog
ctg <- lfa_read_area_as_catalog("SpeciesA", "Area1")
returns_counts_per_tree <- lfa_count_returns_per_tree(ctg)

}
\seealso{
\code{\link{lidR::readLAS}}, \code{\link{lidR::is.empty}},
\code{\link{base::table}}, \code{\link{dplyr::bind_rows}}
}
\keyword{counting}
\keyword{lidR}
\keyword{lidar}
\keyword{returns}
\keyword{tree}
