% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfa_create_boxplots.R
\name{lfa_create_boxplot}
\alias{lfa_create_boxplot}
\title{Create a box plot from a data frame}
\usage{
lfa_create_boxplot(
  data,
  value_column,
  category_column1,
  category_column2,
  title = NULL
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{value_column}{The name of the column containing the values for the box plot.}

\item{category_column1}{The name of the column containing the first categorical variable.}

\item{category_column2}{The name of the column containing the second categorical variable.}

\item{title}{An optional title for the plot. If not provided, a default title is generated based on the data frame name.}
}
\value{
A ggplot object representing the box plot.
}
\description{
This function generates a box plot using ggplot2 based on the specified data frame and columns.
}
\details{
The function creates a box plot where the x-axis is based on the second categorical variable,
the y-axis is based on the specified value column, and the box plots are colored based on the first
categorical variable. The grouping of box plots is done based on the unique values in the second categorical variable.
}
\examples{
\dontrun{
# Assuming you have a data frame 'your_data' with columns 'value', 'category1', and 'category2'
create_boxplot(your_data, "value", "category1", "category2")
}

}
