% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_density_plots.R
\name{lfa_create_density_plots}
\alias{lfa_create_density_plots}
\title{Create density plots for groups in a data frame}
\usage{
lfa_create_density_plots(
  data,
  value_column,
  category_column1 = "area",
  category_column2 = "specie",
  title = NULL,
  xlims = NULL,
  ylims = NULL
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{value_column}{The name of the column containing the values for the density plot.}

\item{category_column1}{The name of the column containing the categorical variable for grouping.}

\item{category_column2}{The name of the column containing the categorical variable for arranging plots.}

\item{title}{An optional title for the plot. If not provided, a default title is generated based on the data frame name.}

\item{xlims}{Optional limits for the x-axis. Should be a numeric vector with two elements (lower and upper bounds).}

\item{ylims}{Optional limits for the y-axis. Should be a numeric vector with two elements (lower and upper bounds).}
}
\value{
A ggplot object representing the density plots arranged in a 2x2 grid.
}
\description{
This function generates density plots using ggplot2 based on the specified data frame and columns.
}
\details{
The function creates density plots where the x-axis is based on the specified value column,
and the density plots are colored based on the first categorical variable. The arrangement of plots
is done based on the unique values in the second categorical variable. The plots are arranged in a 2x2 grid.
}
\examples{
\dontrun{
# Assuming you have a data frame 'your_data' with columns 'value', 'category1', and 'category2'
create_density_plots(your_data, "value", "category1", "category2", title = "Density Plots", xlims = c(0, 10), ylims = c(0, 0.5))
}

}
