% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar_plot.R
\name{lfa_create_grouped_bar_plot}
\alias{lfa_create_grouped_bar_plot}
\title{Create a barplot using ggplot2}
\usage{
lfa_create_grouped_bar_plot(data, grouping_var, value_col, label_col)
}
\arguments{
\item{df}{A data frame containing the relevant columns for the barplot.}

\item{value_column}{The column containing the values to be plotted.}

\item{label_column}{The column used for labeling the bars on the x-axis. Default is "name".}

\item{grouping_column}{The column used for grouping the bars. Default is "species".}
}
\value{
A ggplot2 barplot.
}
\description{
This function generates a barplot using ggplot2 based on the specified data frame columns.
The barplot displays the values from the specified column, grouped by another column.
The grouping can be further differentiated by color if desired.
}
\examples{
# Assuming you have a data frame 'your_data_frame' with columns "name", "species", and "value"
lfa_create_barplot(your_data_frame, value_column = "value", label_column = "name", grouping_column = "species")

}
