% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_per_area.R
\name{lfa_create_plot_per_area}
\alias{lfa_create_plot_per_area}
\title{Create a line plot per area with one color per specie}
\usage{
lfa_create_plot_per_area(data)
}
\arguments{
\item{data}{A data frame with numeric columns and a column named 'specie' for species information.}
}
\value{
A ggplot2 line plot.
}
\description{
This function takes a data frame containing numeric columns and creates a line plot
using ggplot2. Each line in the plot represents a different area, with one color per specie.
}
\examples{
data <- data.frame(
  specie = rep(c("Species1", "Species2", "Species3"), each = 10),
  column1 = rnorm(30),
  column2 = rnorm(30),
  column3 = rnorm(30)
)
lfa_create_plot_per_area(data)

}
