% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ppp_from_area.R
\name{lfa_create_ppp_from_area}
\alias{lfa_create_ppp_from_area}
\title{Create a point pattern from tree detections in a specified area for a given species.}
\usage{
lfa_create_ppp_from_area(species_identifier, area_identifier)
}
\arguments{
\item{species_identifier}{A character string specifying the target species for which
the point pattern is to be generated.}

\item{area_identifier}{A character string specifying the target area for which
the point pattern is to be generated.}
}
\value{
A point pattern representing tree detections for the specified species
within the defined area.
}
\description{
This function generates a point pattern from tree detections for a specific species
within a defined area. It filters the detections using the provided \code{species_identifier}
and \code{area_identifier} parameters. The area is defined by a shapefile named
"research_areas.shp," and the resulting point pattern is created within the specified area.
}
\examples{
lfa_create_ppp_from_area(species_identifier = "SpeciesA", area_identifier = "Area1")

# Create a point pattern for a specific species in a given area
pp <- lfa_create_ppp_from_area(species_identifier = "SpeciesA", area_identifier = "Area1")

}
\seealso{
\code{\link{lfa_get_detections}}, \code{\link{sf::st_transform}},
\code{\link{sf::st_union}}, \code{\link{spatstat.geom::as.owin}},
\code{\link{spatstat.geom::as.ppp}}
}
\keyword{data}
\keyword{detection}
\keyword{pattern}
\keyword{point}
\keyword{spatial}
