% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacked_distributions_plot.R
\name{lfa_create_stacked_distributions_plot}
\alias{lfa_create_stacked_distributions_plot}
\title{Create a stacked distribution plot for tree detections, visualizing the distribution
of a specified variable on the x-axis, differentiated by another variable.}
\usage{
lfa_create_stacked_distributions_plot(
  trees,
  x_value,
  fill_value,
  bin = 100,
  ylab = "Amount trees",
  xlim = c(0, 100),
  ylim = c(0, 1000),
  title =
    "Histograms of height distributions between species 'beech', 'oak', 'pine' and 'spruce' divided by the different areas of Interest"
)
}
\arguments{
\item{trees}{A data frame containing tree detection data.}

\item{x_value}{A character string specifying the column name used for finding the
values on the x-axis of the histogram.}

\item{fill_value}{A character string specifying the column name by which the data
are differentiated in the plot.}

\item{bin}{An integer specifying the number of bins for the histogram. Default is 100.}

\item{ylab}{A character string specifying the y-axis label. Default is "Amount trees."}

\item{xlim}{A numeric vector of length 2 specifying the x-axis limits. Default is c(0, 100).}

\item{ylim}{A numeric vector of length 2 specifying the y-axis limits. Default is c(0, 1000).}

\item{title}{The title of the plot.}
}
\value{
A ggplot object representing the stacked distribution plot.
}
\description{
This function generates a stacked distribution plot using the ggplot2 package,
providing a visual representation of the distribution of a specified variable
(\code{x_value}) on the x-axis, with differentiation based on another variable
(\code{fill_value}). The data for the plot are derived from the provided \code{trees}
data frame.
}
\examples{
# Create a stacked distribution plot for variable "Z," differentiated by "area"
trees <- lfa_get_detections()
lfa_create_stacked_distributions_plot(trees, "Z", "area")

}
\seealso{
\code{\link{ggplot2::geom_histogram}}, \code{\link{ggplot2::facet_wrap}},
\code{\link{ggplot2::ylab}}, \code{\link{ggplot2::scale_fill_brewer}},
\code{\link{ggplot2::coord_cartesian}}
}
\keyword{data}
\keyword{distribution}
\keyword{ggplot2}
\keyword{plot}
\keyword{visualization}
