% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacked_histogram.R
\name{lfa_create_stacked_histogram}
\alias{lfa_create_stacked_histogram}
\title{Create a stacked histogram for tree detections, summing up the values for each species.}
\usage{
lfa_create_stacked_histogram(
  trees,
  x_value,
  fill_value,
  bin = 30,
  ylab = "Frequency",
  xlim = c(0, 100),
  ylim = NULL
)
}
\arguments{
\item{trees}{A data frame containing tree detection data.}

\item{x_value}{A character string specifying the column name used for finding the
values on the x-axis of the histogram.}

\item{fill_value}{A character string specifying the column name by which the data
are differentiated in the plot.}

\item{bin}{An integer specifying the number of bins for the histogram. Default is 30.}

\item{ylab}{A character string specifying the y-axis label. Default is "Frequency."}

\item{xlim}{A numeric vector of length 2 specifying the x-axis limits. Default is c(0, 100).}

\item{ylim}{A numeric vector of length 2 specifying the y-axis limits. Default is NULL.}
}
\value{
A ggplot object representing the stacked histogram.
}
\description{
This function generates a stacked histogram using the ggplot2 package,
summing up the values for each species and visualizing the distribution of
a specified variable (\code{x_value}) on the x-axis, differentiated by another
variable (\code{fill_value}). The data for the plot are derived from the provided
\code{trees} data frame.
}
\examples{
# Create a stacked histogram for variable "Z," differentiated by "area"
trees <- lfa_get_detections()
lfa_create_stacked_histogram(trees, "Z", "area")

}
\seealso{
\code{\link{ggplot2::geom_histogram}}, \code{\link{ggplot2::ylab}},
\code{\link{ggplot2::scale_fill_brewer}}, \code{\link{ggplot2::coord_cartesian}}
}
\keyword{data}
\keyword{distribution}
\keyword{ggplot2}
\keyword{plot}
\keyword{visualization}
