% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tile_locations.R
\name{lfa_create_tile_location_objects}
\alias{lfa_create_tile_location_objects}
\title{Create tile location objects}
\usage{
lfa_create_tile_location_objects()
}
\value{
A vector containing tile location objects.
}
\description{
This function traverses a directory structure to find LAZ files and creates
tile location objects for each file. The function looks into the the \code{data}
directory of the repository/working directory. It then creates \code{tile_location}
objects based on the folder structure. The folder structure should not be
touched by hand, but created by \code{lfa_init_data_structure()} which builds the
structure based on a shape file.
}
\examples{
lfa_create_tile_location_objects()

\dontrun{
lfa_create_tile_location_objects()
}
}
\seealso{
\code{\link{tile_location}}
}
\author{
Jakob Danel
}
