% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{lfa_detection}
\alias{lfa_detection}
\title{Perform tree detection on a lidar catalog and optionally save the results to a file.}
\usage{
lfa_detection(catalog, tile_location, write_to_file = TRUE)
}
\arguments{
\item{catalog}{A lidar catalog containing point cloud data. If set to NULL, the function attempts to read the catalog from the specified tile location.}

\item{tile_location}{An object specifying the location of the lidar tile. If catalog is NULL, the function attempts to read the catalog from this tile location.}

\item{write_to_file}{A logical value indicating whether to save the detected tree information to a file. Default is TRUE.}
}
\value{
A sf style data frame containing information about the detected trees.
}
\description{
This function utilizes lidar data to detect trees within a specified catalog. The detected tree information can be optionally saved to a file in the GeoPackage format. The function uses parallel processing to enhance efficiency.
}
\examples{
# Perform tree detection on a catalog and save the results to a file
lfa_detection(catalog = my_catalog, tile_location = my_tile_location, write_to_file = TRUE)
}
