% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{lfa_download}
\alias{lfa_download}
\title{Download an las file from the state NRW from a specific location}
\usage{
lfa_download(species, name, location)
}
\arguments{
\item{species}{The species of the tree which is observed at this location}

\item{name}{The name of the area that is observed}

\item{location}{An sf object, which holds the location information for the area where the tile should be downloaded from.}
}
\value{
The LASCatalog object of the downloaded file
}
\description{
It will download the file and save it to data/\if{html}{\out{<species>}}/\if{html}{\out{<name>}} with the name of the tile
}
