% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_areas_sf.R
\name{lfa_download_areas}
\alias{lfa_download_areas}
\title{Download areas based on spatial features}
\usage{
lfa_download_areas(sf_areas)
}
\arguments{
\item{sf_areas}{Spatial features representing areas to be downloaded.
It must include columns like "species" "name"
See details for more information.}
}
\value{
None
}
\description{
This function initiates the data structure and downloads areas based on spatial features.
}
\details{
The input data frame, \code{sf_areas}, must have the following columns:
\itemize{
\item "species": The species associated with the area.
\item "name": The name of the area.
}

The function uses the \code{lfa_init_data_structure} function to set up the data structure
and then iterates through the rows of \code{sf_areas} to download each specified area.
}
\examples{
\dontrun{
lfa_download_areas(sf_areas)
}


\dontrun{
# Example spatial features data frame
sf_areas <- data.frame(
  species = c("SpeciesA", "SpeciesB"),
  name = c("Area1", "Area2"),
  # Must include also other attributes specialized to sf objects
  # such as geometry, for processing of the download
)

lfa_download_areas(sf_areas)
}
}
\author{
Jakob Danel
}
