% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_n_nearest_trees.R
\name{lfa_find_n_nearest_trees}
\alias{lfa_find_n_nearest_trees}
\title{Find n Nearest Trees}
\usage{
lfa_find_n_nearest_trees(trees, n = 100)
}
\arguments{
\item{trees}{A sf object containing tree coordinates.}

\item{n}{The number of nearest trees to find for each tree (default is 100).}
}
\value{
A data frame with additional columns representing the distances to the n nearest trees.
}
\description{
This function calculates the distances to the n nearest trees for each tree in the input dataset.
}
\examples{
# Load tree data using lfa_get_detections() (not provided)
tree_data <- lfa_get_detections()

# Filter tree data for a specific species and area
tree_data = tree_data[tree_data$specie == "pine" & tree_data$area == "greffen", ]

# Find the 100 nearest trees for each tree in the filtered dataset
tree_data <- lfa_find_n_nearest_trees(tree_data)

}
