% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_areas.R
\name{lfa_get_all_areas}
\alias{lfa_get_all_areas}
\title{Retrieve a data frame containing all species and corresponding areas.}
\usage{
lfa_get_all_areas()
}
\value{
A data frame with columns "specie" and "area" containing information about
all species and their associated areas.
}
\description{
This function scans the "data" directory within the current working directory to
obtain a list of species. It then iterates through each species to retrieve the list
of areas associated with that species. The resulting data frame contains two columns:
"specie" representing the species and "area" representing the corresponding area.
}
\examples{
# Retrieve a data frame with information about all species and areas
all_areas_df <- lfa_get_all_areas()

}
\seealso{
\code{\link{list.dirs}}
}
\keyword{data}
\keyword{file}
\keyword{manipulation}
\keyword{system}
