% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detection_area.R
\name{lfa_get_detection_area}
\alias{lfa_get_detection_area}
\title{Get Detection for an area}
\usage{
lfa_get_detection_area(species, name)
}
\arguments{
\item{species}{A character string specifying the target species.}

\item{name}{A character string specifying the name of the tile.}
}
\value{
A Simple Features (SF) data frame containing tree detection information for the specified species and tile.
}
\description{
Retrieves the tree detection information for a specified species and tile.
}
\details{
This function reads tree detection data from geopackage files within the specified tile location for a given species. It then combines the data into a single SF data frame and returns it. The function assumes that the tree detection files follow a naming convention with the pattern "_detection.gpkg".
}
\examples{
# Retrieve tree detection data for species "example_species" in tile "example_tile"
trees_data <- lfa_get_detection_tile_location("example_species", "example_tile")

# Example usage:
trees_data <- lfa_get_detection_tile_location("example_species", "example_tile")

# No trees found scenario:
empty_data <- lfa_get_detection_tile_location("nonexistent_species", "nonexistent_tile")
# The result will be an empty data frame if no trees are found for the specified species and tile.

# Error handling:
# In case of invalid inputs, the function may throw errors. Ensure correct species and tile names are provided.

}
\references{
This function is part of the LiDAR Forest Analysis (LFA) package.
}
\seealso{
\code{\link{get_tile_dir}}
}
\keyword{spatial}
