% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{lfa_get_detections}
\alias{lfa_get_detections}
\title{Retrieve aggregated detection data for multiple species.}
\usage{
lfa_get_detections()
}
\value{
A data frame containing aggregated detection data for multiple species.
}
\description{
This function obtains aggregated detection data for multiple species by iterating
through the list of species obtained from \code{\link{lfa_get_species}}. For each
species, it calls \code{\link{lfa_get_detections_species}} to retrieve the
corresponding detection data and aggregates the results into a single data frame.
The resulting data frame includes columns for the species, tree detection data,
and the area in which the detections occurred.
}
\examples{
lfa_get_detections()

# Retrieve aggregated detection data for multiple species
detections_data <- lfa_get_detections()

}
\seealso{
\code{\link{lfa_get_species}}, \code{\link{lfa_get_detections_species}}

Other data retrieval functions: 
\code{\link{lfa_get_species}()}
}
\concept{data retrieval functions}
\keyword{aggregation}
\keyword{data}
\keyword{manipulation}
