% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{lfa_get_species}
\alias{lfa_get_species}
\title{Get a list of species from the data directory.}
\usage{
lfa_get_species()
}
\value{
A character vector containing the names of species found in the "data" directory.
}
\description{
This function retrieves a list of species by scanning the "data" directory
located in the current working directory.
}
\examples{
# Retrieve the list of species
species_list <- lfa_get_species()

}
\references{
This function relies on the \code{\link{list.dirs}} function for directory listing.
}
\seealso{
\code{\link{list.dirs}}

Other data retrieval functions: 
\code{\link{lfa_get_detections}()}
}
\concept{data retrieval functions}
\keyword{data}
\keyword{file}
\keyword{manipulation}
\keyword{system}
