% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_correction.R
\name{lfa_ground_correction}
\alias{lfa_ground_correction}
\title{Correct the point clouds for correct ground imagery}
\usage{
lfa_ground_correction(ctg, tile_location)
}
\arguments{
\item{ctg}{An LASCatalog object. If not null, it will perform the actions on this object, if NULL
inferring the catalog from the tile_location}

\item{tile_location}{A tile_location type object holding the information about the location of the
cataog. This is used to save the catalog after processing too.}
}
\value{
A catalog with the corrected z values. The catalog is always stored at tile_location and
holding only the transformed values.
}
\description{
This function is needed to correct the Z value of the point cloud, relative to the real
ground height. After using this function to your catalog, the Z values can be seen as the
real elevation about the ground. At the moment the function uses the \code{tin()} function from
the \code{lidr} package. \emph{NOTE}: The operation is inplace and can not be reverted, the old values
of the point cloud will be deleted!
}
\author{
Jakob Danel
}
