% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{lfa_init}
\alias{lfa_init}
\title{Initialize LFA (LiDAR forest analysis) data processing}
\usage{
lfa_init(sf_file = "research_areas.shp")
}
\arguments{
\item{sf_file}{A character string specifying the path to the shapefile containing
spatial features of research areas.}
}
\value{
A vector containing tile location objects.
}
\description{
This function initializes the LFA data processing by reading a shapefile containing
spatial features of research areas, downloading the specified areas, and creating
tile location objects for each area.
}
\details{
This function reads a shapefile (\code{sf_file}) using the \code{sf} package, which should
contain information about research areas. It then calls the \code{lfa_download_areas}
function to download the specified areas and \code{lfa_create_tile_location_objects}
to create tile location objects based on Lidar data files in those areas. The
shapefile \emph{MUST} follow the following requirements:
\itemize{
\item Each geometry must be a single object of type polygon
\item Each entry must have the following attributes:
\itemize{
\item species: A string describing the tree species of the area.
\item name: A string describing the location of the area.
}
}
}
\examples{
\dontrun{
# Initialize LFA processing with the default shapefile
lfa_init()

# Initialize LFA processing with a custom shapefile
lfa_init("custom_areas.shp")
}

\dontrun{
# Example usage with the default shapefile
lfa_init()

# Example usage with a custom shapefile
lfa_init("custom_areas.shp")
}

}
