% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data_structure.R
\name{lfa_init_data_structure}
\alias{lfa_init_data_structure}
\title{Initialize data structure for species and areas}
\usage{
lfa_init_data_structure(sf_species)
}
\arguments{
\item{sf_species}{A data frame with information about species and associated areas.}
}
\value{
None
}
\description{
This function initializes the data structure for storing species and associated areas.
}
\details{
The input data frame, \code{sf_species}, should have at least the following columns:
\itemize{
\item "species": The names of the species for which the data structure needs to be initialized.
\item "name": The names of the associated areas.
}

The function creates directories based on the species and area information provided in
the \code{sf_species} data frame. It checks whether the directories already exist and creates
them if they don't.
}
\examples{
\dontrun{
# Example species data frame
sf_species <- data.frame(
  species = c("SpeciesA", "SpeciesB"),
  name = c("Area1", "Area2"),
  # Other necessary columns
)

lfa_init_data_structure(sf_species)
}

\dontrun{
# Example species data frame
sf_species <- data.frame(
  species = c("SpeciesA", "SpeciesB"),
  name = c("Area1", "Area2"),
  # Other necessary columns
)

lfa_init_data_structure(sf_species)
}

}
