% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsd_from_vec.R
\name{lfa_jsd_from_vec}
\alias{lfa_jsd_from_vec}
\title{Compute Jensen-Shannon Divergence from Vectors}
\usage{
lfa_jsd_from_vec(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
Jensen-Shannon Divergence between the density distributions of x and y.
}
\description{
This function calculates the Jensen-Shannon Divergence (JSD) between two vectors.
}
\examples{
x <- rnorm(100)
y <- rnorm(100, mean = 2)
lfa_jsd_from_vec(x, y)

}
