% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld_from_vec.R
\name{lfa_kld_from_vec}
\alias{lfa_kld_from_vec}
\title{Compute Kullback-Leibler Divergence from Vectors}
\usage{
lfa_kld_from_vec(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
Kullback-Leibler Divergence between the density distributions of x and y.
}
\description{
This function calculates the Kullback-Leibler Divergence (KLD) between two vectors.
}
\examples{
x <- rnorm(100)
y <- rnorm(100, mean = 2)
lfa_kld_from_vec(x, y)

}
