% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{lfa_ks_test}
\alias{lfa_ks_test}
\title{Kolmogorov-Smirnov Test Wrapper Function}
\usage{
lfa_ks_test(x, y, output_variable = "p.value", ...)
}
\arguments{
\item{x}{A numeric vector representing the first sample.}

\item{y}{A numeric vector representing the second sample.}

\item{output_variable}{A character string specifying the output variable to extract from the ks.test result.
Default is "p.value". Other possible values include "statistic" and "alternative".}

\item{...}{Additional arguments to be passed to the ks.test function.}
}
\value{
A numeric value representing the specified output variable from the KS test result.
}
\description{
This function serves as a wrapper for the Kolmogorov-Smirnov (KS) test between two samples.
}
\details{
The function uses the ks.test function to perform a two-sample KS test and returns the specified output variable.
The default output variable is the p-value. Other possible output variables include "statistic" and "alternative".
}
\examples{
# Perform KS test and extract the p-value
result <- lfa_ks_test(sample1, sample2)
print(result)

# Perform KS test and extract the test statistic
result_statistic <- lfa_ks_test(sample1, sample2, output_variable = "statistic")
print(result_statistic)

}
\seealso{
\code{\link{ks.test}}
}
