% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{lfa_load_ctg_if_not_present}
\alias{lfa_load_ctg_if_not_present}
\title{Loading the catalog if it is not present}
\usage{
lfa_load_ctg_if_not_present(ctg, tile_location)
}
\arguments{
\item{ctg}{Catalog object. Can be NULL}

\item{tile_location}{The location to look for the catalog tiles, if their
are not present}
}
\value{
The provided ctg object if not null, else the catalog for the tiles
of the tile_location.
}
\description{
This function checks if the catalog is \code{NULL}. If it is it will load the
catalog from the \code{tile_location}
}
