% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion_matrix.R
\name{lfa_plot_confusion_matrix}
\alias{lfa_plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
lfa_plot_confusion_matrix(conf_matrix)
}
\arguments{
\item{conf_matrix}{Confusion matrix, typically obtained from classification evaluation.}
}
\value{
A ggplot object representing the confusion matrix heatmap plot.
}
\description{
This function generates a heatmap plot of a confusion matrix using ggplot2.
}
\details{
The function takes a confusion matrix as input and generates a heatmap plot using ggplot2.
The plot represents the relationship between the predicted and actual classes, with cell colors indicating the frequency of each combination.
Additionally, the plot includes labels for accuracy and kappa statistics based on the confusion matrix.
}
\examples{
# Example confusion matrix
cm <- table(predicted = c("A", "B", "A", "B"), actual = c("A", "A", "B", "B"))
# Plot confusion matrix
lfa_plot_confusion_matrix(cm)

}
