% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{lfa_random_forest}
\alias{lfa_random_forest}
\title{Random Forest Classifier with Leave-One-Out Cross-Validation}
\usage{
lfa_random_forest(
  tree_data,
  excluded_input_columns,
  response_variable = "specie",
  ntree = 100,
  seed = 123,
  ...
)
}
\arguments{
\item{tree_data}{A data frame containing the tree data, including the response variable ("specie") and predictor variables.}

\item{excluded_input_columns}{A character vector specifying columns to be excluded from predictor variables.}

\item{response_variable}{The response variable to be predicted (default is "specie").}

\item{seed}{An integer to set the seed for reproducibility (default is 123).}

\item{...}{Additional parameters to be passed to the randomForest function.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{predicted_species_absolute}: A data frame with observed and predicted species for each area.
\item \code{predicted_species_relative}: A data frame wit the relative precictions per speices and areas, normalized by the total predictions in each area.
\item \code{confusion_matrix}: A confusion matrix showing the counts of predicted vs. observed species.
\item \code{accuracy}: The accuracy of the model, calculated as the sum of diagonal elements in the confusion matrix divided by the total count.
\item \code{formula}: The formula used for modeling.
}
}
\description{
This function performs a random forest classification using leave-one-out cross-validation for each area in the input tree data.
It returns a list containing various results, including predicted species, confusion matrix, accuracy, and the formula used for modeling.
}
\examples{
# Assuming tree_data is defined
results <- lfa_random_forest(tree_data, excluded_input_columns = c("column1", "column2"))

# Print the list of results
print(results)

}
