% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_chunk.R
\name{lfa_rasterize_chunk}
\alias{lfa_rasterize_chunk}
\title{Rasterize Lidar Chunk}
\usage{
lfa_rasterize_chunk(chunk, ...)
}
\arguments{
\item{chunk}{Lidar chunk object to be rasterized.}

\item{...}{Additional arguments to be passed to the underlying lidR::rasterize_canopy function.}
}
\value{
A raster layer representing the rasterized canopy.
}
\description{
This function rasterizes a Lidar chunk to generate a raster representation of the canopy.
}
\details{
The function takes a Lidar chunk as input and uses lidR::rasterize_canopy to generate a raster representation of the canopy.
Additional arguments can be passed to customize the rasterization process.
}
\examples{
# Example Lidar chunk
lidar_chunk <- readLAS("lidar_data.las", select = "xyz")
# Rasterize Lidar chunk
rasterized_canopy <- lfa_rasterize_chunk(lidar_chunk)

}
