% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_to_md.R
\name{lfa_rd_to_qmd}
\alias{lfa_rd_to_qmd}
\title{Convert Rd File to Markdown}
\usage{
lfa_rd_to_qmd(rdfile, outfile, append = FALSE)
}
\arguments{
\item{rdfile}{The path to the Rd file or a parsed Rd object.}

\item{outfile}{The path to the output Markdown file (including the file extension).}

\item{append}{Logical, indicating whether to append to an existing file (default is FALSE).}
}
\value{
This function does not explicitly return any value. It saves the converted Markdown file
at the specified location as described in the details section.
}
\description{
\emph{IMPORTANT NOTE:}
This function is nearly identical to the \code{Rd2md::Rd2markdown} function from the \code{Rd2md}
package. We needed to implement our own version of it because of various reasons:
\itemize{
\item The algorithm uses hardcoded header sizes (h1 and h2 in original) which is not feasible for our use-case of the markdown.
\item We needed to add some Quarto Markdown specifics, e.g. to make sure that the examples will not be runned.
\item We want to exclude certain tags from our implementation.
}
}
\details{
For that reason we copied the method and made changes as needed and also added this custom documentation.

This function converts an Rd (R documentation) file to Markdown format (.md) and
saves the converted file at the specified location. The function allows appending
to an existing file or creating a new one. The resulting Markdown file includes
sections for the function's name, title, and additional content such as examples,
usage, arguments, and other sections present in the Rd file.

The function performs the following steps:
\itemize{
\item Parses the Rd file using the Rd2md package.
\item Creates a Markdown file with sections for the function's name, title, and additional content.
\item Appends the content to an existing file if \code{append} is set to TRUE.
\item Saves the resulting Markdown file at the specified location.
}
}
\examples{
# Convert Rd file to Markdown and save it
lfa_rd_to_md("path/to/your/file.Rd", "path/to/your/output/file.md")

# Convert Rd file to Markdown and append to an existing file
lfa_rd_to_md("path/to/your/file.Rd", "path/to/existing/output/file.md", append = TRUE)

}
\seealso{
\code{\link{Rd2md::parseRd}}
}
