% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_to_results.R
\name{lfa_rd_to_results}
\alias{lfa_rd_to_results}
\title{Convert Rd Files to Markdown and Merge Results}
\usage{
lfa_rd_to_results()
}
\value{
This function does not explicitly return any value. It performs the conversion,
merging, and saving operations as described in the details section.
}
\description{
This function converts all Rd (R documentation) files in the "man" directory
to Markdown format (.qmd) and saves the converted files in the "results/appendix/package-docs" directory.
It then merges the converted Markdown files into a single string and saves
the merged content into a file named "docs.qmd" in the "results/appendix/package-docs" directory.
}
\details{
The function performs the following steps:
\itemize{
\item Removes any existing "docs.qmd" file in the "results/appendix/package-docs" directory.
\item Finds all Rd files in the "man" directory.
\item Converts each Rd file to Markdown format (.qmd) using the \code{lfa_rd_to_qmd} function.
\item Saves the converted Markdown files in the "results/appendix/package-docs" directory.
\item Merges the content of all converted Markdown files into a single string.
\item Saves the merged content into a file named "docs.qmd" in the "results/appendix/package-docs" directory.
}
}
\examples{
# Convert Rd files to Markdown and merge the results
lfa_rd_to_results()

}
\seealso{
\code{\link{lfa_rd_to_qmd}}, \code{\link{lfa_merge_and_save}}
}
