% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_catalog_per_area.R
\name{lfa_read_area_as_catalog}
\alias{lfa_read_area_as_catalog}
\title{Read LiDAR data from a specified species and location as a catalog.}
\usage{
lfa_read_area_as_catalog(specie, location_name)
}
\arguments{
\item{specie}{A character string specifying the species of interest.}

\item{location_name}{A character string specifying the name of the location.}
}
\value{
A \code{lidR::LAScatalog} object containing the LiDAR data from the specified location and species.
}
\description{
This function constructs the file path based on the specified \code{specie} and \code{location_name},
lists the directories at that path, and reads the LiDAR data into a \code{lidR::LAScatalog}.
}
\examples{
\dontrun{
lfa_read_area_as_catalog("beech", "location1")
}
}
