% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_per_class.R
\name{lfa_recall_per_class}
\alias{lfa_recall_per_class}
\title{Calculate Recall per Class from Confusion Matrix}
\usage{
lfa_recall_per_class(confusion_matrix)
}
\arguments{
\item{confusion_matrix}{Confusion matrix obtained from a classification evaluation.}
}
\value{
A numeric vector representing recall for each class.
}
\description{
This function calculates recall for each class based on the provided confusion matrix.
}
\details{
Recall (Sensitivity or True Positive Rate) is a measure of the ability of a classification model to identify all relevant instances. It is calculated as the ratio of true positives to the sum of true positives and false negatives.
}
\examples{
# Example confusion matrix
cm <- table(predicted = c("A", "B", "A", "B"), actual = c("A", "A", "B", "B"))
# Calculate recall per class
recall_vector <- lfa_recall_per_class(cm)

}
\seealso{
\code{\link{lfa_precision_per_class}}
}
