% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test_symmetric.R
\name{lfa_run_test_symmetric}
\alias{lfa_run_test_symmetric}
\title{Symmetric Pairwise Test for Categories}
\usage{
lfa_run_test_symmetric(data, data_column, category_column, test_function, ...)
}
\arguments{
\item{data}{A data frame containing the relevant columns.}

\item{data_column}{A character string specifying the column containing the numerical data.}

\item{category_column}{A character string specifying the column containing the categorical variable.}

\item{test_function}{A function used to perform the pairwise test between two sets of data.
It should accept two vectors of numeric data and additional parameters specified by \code{...}.
The function should return a numeric value representing the test result.}

\item{...}{Additional parameters to be passed to the \code{test_function}.}
}
\value{
A data frame representing the results of the symmetric pairwise tests between categories.
}
\description{
This function performs a symmetric pairwise test for categories using a user-defined \code{test_function}.
}
\details{
The function calculates the test results for each unique combination of categories using the specified
\code{test_function}. The resulting table is symmetric, containing the test results for comparisons
from the rows to the columns. The upper triangle of the matrix is filled with \code{NA} to avoid duplicate results.
}
\examples{
# Define a custom test function
custom_test_function <- function(x, y) {
  # Your test logic here
  # Return a numeric result
  return(mean(x) - mean(y))
}

# Perform a symmetric pairwise test
result <- lfa_run_test_symmetric(your_data, "numeric_column", "category_column", custom_test_function)

}
\seealso{
\code{\link{outer}}, \code{\link{Vectorize}}
}
