% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all_neighbours.R
\name{lfa_save_all_neighbours}
\alias{lfa_save_all_neighbours}
\title{Save Neighbors for All Areas}
\usage{
lfa_save_all_neighbours(n = 100)
}
\arguments{
\item{n}{The number of nearest trees to find for each tree (default is 100).}
}
\description{
This function iterates through all detection areas, finds the n nearest trees for each tree,
and saves the result to a GeoPackage file for each area.
}
\examples{
# Save neighbors for all areas with default value (n=100)
lfa_save_all_neighbours()

# Save neighbors for all areas with a specific value of n (e.g., n=50)
lfa_save_all_neighbours(n = 50)
}
