% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{lfa_segmentation}
\alias{lfa_segmentation}
\title{Segment the elements of an point cloud by trees}
\usage{
lfa_segmentation(ctg, tile_location)
}
\arguments{
\item{ctg}{An LASCatalog object. If not null, it will perform the actions on this object, if NULL
inferring the catalog from the tile_location}

\item{tile_location}{A tile_location type object holding the information about the location of the
catalog. This is used to save the catalog after processing too.}
}
\value{
A catalog where each chunk has additional \code{treeID} values indicating the belonging tree.
}
\description{
This function will try to to divide the hole point cloud into unique trees.
Therefore it is assigning for each chunk of the catalog  a \code{treeID} for each
point. Therefore the algorithm uses the \code{li2012} implementation with the
following parameters: \code{li2012(dt1 = 2, dt2 = 3, R = 2, Zu = 10, hmin = 5, speed_up = 12)}
\emph{NOTE}: The operation is in place and can not be reverted, the old values
of the point cloud will be deleted!
}
\author{
Jakob Danel
}
