% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_flags.R
\name{lfa_set_flag}
\alias{lfa_set_flag}
\title{Set a flag to indicate the completion of a specific process.}
\usage{
lfa_set_flag(flag_name)
}
\arguments{
\item{flag_name}{A character string specifying the name of the flag file. It should be a descriptive and unique identifier for the process being flagged.}
}
\value{
This function does not have a formal return value.
}
\description{
This function creates a hidden flag file at a specified location within the working directory to indicate that a particular processing step has been completed. If the flag file already exists, a warning is issued.
}
\examples{
# Set the flag for a process named "data_processing"
lfa_set_flag("data_processing")
}
