% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_all_areas.R
\name{lfa_visit_all_areas}
\alias{lfa_visit_all_areas}
\title{Visit All Areas and Apply Preprocessing Function}
\usage{
lfa_visit_all_areas(preprocessing_function, areas = lfa_get_all_areas(), ...)
}
\arguments{
\item{preprocessing_function}{The preprocessing function to be applied to each area. It should take specie, area, and additional parameters as inputs.}

\item{areas}{Data frame containing information about different areas, including columns "specie" and "area."}

\item{...}{Additional arguments to be passed to the preprocessing function.}
}
\value{
A list containing the results of applying the preprocessing function to each area.
}
\description{
This function iterates over all specified areas and applies a preprocessing function to each one.
}
\details{
The function iterates over all areas specified in the 'areas' parameter, and for each area, it applies the provided preprocessing function.
The 'areas' parameter is expected to be a data frame with columns "specie" and "area," containing information about different areas to visit.
Additional arguments passed via '...' are forwarded to the preprocessing function.
}
\examples{
# Example preprocessing function
my_preprocessing_function <- function(specie, area, ...) {
  # Your preprocessing logic here
  # Return the result
  return(result)
}
# Visit all areas and apply the preprocessing function
results_list <- lfa_visit_all_areas(my_preprocessing_function)

}
\seealso{
\code{\link{lfa_get_all_areas}}
}
