% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_rf_metrics.R
\name{lfa_visualize_rf_metrics}
\alias{lfa_visualize_rf_metrics}
\title{Visualize Precision and Recall Metrics for Random Forest Classification}
\usage{
lfa_visualize_rf_metrics(metrics_df)
}
\arguments{
\item{metrics_df}{Data frame containing precision and recall metrics for each class.}
}
\value{
A ggplot object representing the bar plot of precision and recall metrics.
}
\description{
This function creates a bar plot to visualize precision and recall metrics for each class obtained from a Random Forest classification.
}
\details{
The function creates a bar plot to visualize precision and recall metrics for each class obtained from a Random Forest classification.
}
\examples{
# Example data frame containing precision and recall metrics
example_metrics_df <- data.frame(
  Class = c("ClassA", "ClassB"),
  Precision = c(0.85, 0.92),
  Recall = c(0.78, 0.88)
)
# Visualize precision and recall metrics
lfa_visualize_rf_metrics(example_metrics_df)

}
\seealso{
\code{\link{lfa_calculate_rf_metrics}}
}
